Fix #62
Fixes an error where some p-values are saved as `str` by checking if they can be cast to `float` then casting if so. If not, a `ValueError` is raised. This is probably not the best solution ("why are they `str` in the first place?") but it fixes the issue in a semi-sane way.